Rename ekuad to eko
