+++ crates/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 
-members = ["eko", "ekore", "ekuad"]
+members = ["eko", "ekore"]
