+++ crates/eko/Cargo.toml
@@ -3,6 +3,9 @@ name = "eko"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "ekors"
+crate-type = ["cdylib"]
 
 [dependencies]
+ekore = { version = "0.1.0", path = "../ekore" }
