+++ crates/eko/pyproject.toml
@@ -1,10 +1,10 @@
 [build-system]
-requires = ["maturin>=0.15,<0.16"]
+requires = ["maturin>=1.1,<2.0"]
 build-backend = "maturin"
 
 [project]
-name = "ekuad"
-requires-python = ">=3.7"
+name = "ekors"
+requires-python = ">=3.8"
 classifiers = [
   "Programming Language :: Rust",
   "Programming Language :: Python :: Implementation :: CPython",
