+++ crates/try.sh
@@ -1,4 +1,4 @@
-pip install -e ekuad
+pip install -e eko
 
 printf "\nTry to use the CFFI package:\n"
 python integrate.py
